moved file utils to root of minds services.
